\name{TF_acc_21_w_0}
\alias{TF_acc_21_w_0}
\docType{data}
\title{
Contains an output of the accumulation function.
}
\description{
\emph{TF_acc_21_w_0} is a list of 4 elements (\emph{acc_vector},
\emph{acc_type}, \emph{chr}, \emph{w}). It is the output of the accumulation
function with acctype="TF", chr="chr21", w=0. \emph{TF_acc_21_w_0} is included
in the \emph{data_man} collection.
}
\usage{data("data_man")}
\value{
None, the function is invoked for its side effect.
}
\format{
An object of class list.
}
\examples{
# loading data
data("data_man")
# TF_acc_21_w_0 is in the data_man collection of datasets
head(TF_acc_21_w_0)
}
\keyword{datasets}
